+++ redisbungee-abstraction-layer-invk-redisvelocity-implementation/src/main/java/de/simonsator/abstractredisbungee/invk/InvvkFakeJedis.java
@@ -1,4 +1,4 @@
-package de.simonsator.abstractredisbungee.legacy;
+package de.simonsator.abstractredisbungee.invk;
 
 import de.simonsator.abstractredisbungee.fakejedis.FakeJedis;
 import de.simonsator.abstractredisbungee.fakejedis.FakeJedisTransaction;
