+++ redisbungee-abstraction-layer-invk-redisvelocity-implementation/src/main/java/de/simonsator/abstractredisbungee/invk/InvvkRedisVelocityAPI.java
@@ -1,4 +1,4 @@
-package de.simonsator.abstractredisbungee.legacy;
+package de.simonsator.abstractredisbungee.invk;
 
 import com.velocitypowered.api.proxy.server.ServerInfo;
 import de.simonsator.abstractredisbungee.FakeRedisVelocityAPI;
